+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 - Fix issue on dataset/widget filter step.
 - Adds support for more datatypes in chart widgets.
+- Add metadata fields to dataset creation and list
+- Improved 404 page
+- Added sitemap.xml and sitemap.html pages
 - Responsive and cross-browser changes
 - Fix issue on placement of content in site's page tree
 - Improve widget rendering performance